log-push-template/exp1/requirements.txt
python-logstash==0.4.8
